Integrated webrtc for true p2p communication
Before all messages were relayed through server instead of true
p2p communication, but with this commit that has changed.

A server is still needed to track which peers are online and
to relay signaling information to/from peers.

Tested with chrome and it works but firefox did not work :(