Added paths to blocks
Before blocks were all stored in a global space, but
now they can have paths.

For example:
`PULL /` will pull all the blocks in the root path.
`PULL /<post_id>/comments` will pull all the blocks that are comments for the post.
`PUSH /<post_id>/comments` will store a comment for the post.

TODO paths are not normalized. So `/`, ``, `/////` are all different
paths when they should be the same.