Channels implement the EventEmitter interface
This gives chanels greater flexiblity in that the nodejs
EventEmitter functionality can be used without reinventing the
wheel. It comes with the downside that typescripts type checking
can't be used on the arguments for 'on' and 'emit'.

As more events get added to Channels, this change will greatly
reduce the boilerplate that would have been required previously.