Refractored channels to improve modularity
Different types of channels have been broken out
into their own file.

Also, the WrtcChannelManager now uses an arbitrary ChannelManager
to relay it's signaling messages. This allows for the underlying
transports to be swapped without the rest of the codebase caring.

Related to #13 because with the modularized code, the webrtc channels
can be swapped out for some other generic channels.