Fixed race condition while starting channel manager
The code that started the channel manager (cm) would
initialize the cm, then ask the server for peers.
This code would not call the callback which indicates
that the cm is ready until _after_ it received the response
from the server.

If a message where to come in before the server's message
was received, the message would be lost. To fix this,
the callback is called after the socketio is setup instead
of after the receiving the 'peer found' message from the socketio
server.